refactor: divide form components into new files
